Pewter Museum & Viridian Forest warps
Required a little tinkering because I'm vain and wanted the door to be accurate. Please understand.

When doing Viridian Forest I had a bit of a moment and kept trying to use the entirely wrong building when it was staring me in the face. I am sorry.

I'm getting an 8-bit error on Mewtwo's object code in Cerulean Cave B1F. Misty may need to check this, I'm just getting these locations put in the overworld.