Gorochu
This is the first use of the placeholder graphic. Works like a charm - copypaste the folder whenever you add something new.

This required me to change the layout.link for cries but it seems to work just fine. Earthtron cry is now available for use on other Pokemon, which I think could be neat for Shi-Shi.