Nihon Pokemon Centre for Silent Hills
This adds the Pokemon Centre infrastructure for Nihon and Silent Hills's one.

For all future Nihon Pokemon Centres, they must refer to the 2F one for the warp, and the map blocks for 1Fs should use the Nihon variant.

Kanto will follow suit (eventually). The aim is to make a more immersive region structure.