Cal and Silver's Houses
This covers everything but Blue's Lab.

Cal's House has Cal play the role of SW97 Gold, with his mother lamenting his leaving and his brother Ken praising his travels.

Silver's House is a carbon copy of the original. It's worth noting that Silver is less of an arsehole in SW97, so things could change wrt dialogue. This also violates some canon: It's heavily implied in HGSS onwards that Ariana is Silver's mother, chiefly because of her hair. It's worth noting that this hasn't been fully substantiated, and in GSC itself, Ariana has brown/purple hair in all interpretations, which is the basis for the theory. This warrants further discussion.