9 more Pokemon!
Toedscool, Toedscruel, Blottle, Pendraken, Kolta, Balumba, Magnetite, Magnezone, and Gavillain.

Balumba has a small chance to hold a Fairy Feather, which makes it vaguely obtainable.

Kolta is now found in the day in the Burned Tower, whereas Houndour is seen at night. Technically balances it, right?

Gavillain is a rare Dark Cave encounter on the Blackthorn side, dragons and all that.

Balumba is a rare encounter in the National Park!