Like 40 new maps
The aim was simply to make the rom build with EVERY currently posited map in it. This adds EVERYTHING.

Ground rules;
- All temp maps are 10x10 cave tileset goons with nothing in them
- Connections are NOT done
- Fishgroups and music are likely incorrect and/or unfinished
- No wild data

This is extremely achievable, as numerous unreferenced maps are unused in the block and script banks. Makes doing this a cinch!

All of this basically means: You can get creative now and start brainstorming what to do with all these maps. Get them fleshed out, connected, and...not being nothing. It's like ordering in a bunch of blank canvasses. Go nuts!