Fossil scientist is now functional tm
There's a couple bugs here and there, but it's good enough and I think the bugs are related to how I tested it.

Doesn't use specials because it seems to just cause comical glitches to occur.

The Museum and Ruins of Alph fossil guys will talk differently, implying they're different individuals. The Museum one talks like in HGSS, the RoA one talks like the RBY one and comments on the Unown researcher's lack of care for fossils.

Also reused some PokeCom centre lore.