PCC enhancements (and GS Ball restoration)
This adds a bunch of PCC-related stuff to the building, most notably the GS Ball, making Celebi obtainable natively.

Chances are, we would repeat this for the MysticTicket and Mew.

This restores a bunch of text and adapts it to various bits and pieces. Of course, a lot still needs to be done. Very happy with how it's turning out, though!