More accurate News Machine
Turns out it's that massive...thing.

The screen is all fuzz, but it seems like it was meant to change.

We ought to include this stuff
https://www.youtube.com/watch?v=CmvFT4PLAMQ&ab_channel=H%C3%A1%C4%8Dky

But it's hard, so I haven't done it (yet).

I don't think people would kill us if it was just this, but with random quotes from the time period, but...