(very rough) Winner's Path
This implements Winner's Path into the game, complete with trainers and wild Pokemon. There's still a lot to do, though.

The Debug Lady that gave Celebi now gives Furret because it gets all the important HMs, which we need to...test things. Naturally, you have all the badges now too!