Unnumbered Route
This route is unnumbered for now, because it's technically the last Nihon Route (numbered 23 in-prototype) and I don't know how to number it yet. It's easily find/replaced once we're sure, anyway.

Wild data uses the prototype's plus a couple edits, levels are kept low as you're actually entering here.

It's official: Nihon is connected to Johto and Kanto now!