Safari Game battle type & bug fix
This adds the restoration of the Safari Zone from Pokemon October. There isn't a proper place for this yet, but there will be soon!

It seems to have issues with tiles, which I am assuming to either be the tilemap clearing or something relating to Zeta's custom ball palettes, since I had to add the Safari Ball. Can you get to that, @ZetaNull?

https://github.com/pokeachromicdevs/pokeoctober/commit/2ff835730b2a61a757ad67f13e50480c965d2a0c

also fixed the unnumbered route errors.