Lots of new trainer classes & pics, port kep cries
This adds Tohjo Giovanni, Archer, Ariana, Prof. Blue, and fixes various bugs with my last commit. Also fixes the bug with Chris's img not having an embedded palette.

I also ported over all the KEP cries that are available as a bunch were incomplete or seemingly being remade.

Next step is to integrate much of this.