Pokemon Pagoda
This gives Sepia City all of its locations.

The pagoda uses references to the Five-Story Pagoda story from 1891, with Sages using various 成语 translated into English.

This also adds the Elder class, assigning it to Li and (now) Roen.

The Elder in the pagoda is not yet battle-able as I haven't really fit anything into the story. No cutscenes, etc, it's all too early.