The Safari Zone is now functional!
This adds a clause into the wild encounter setup that sets the Safari BattleType if you're within the relevant landmarks. Therefore, all that's now needed is a way to check the Safari Balls in the player's bag before kicking them out, which should be easily triggered by looking in their bag.