Summer Beach House (buggy!)
This adds the Summer Beach House from Pokemon Yellow to Route 19, in an effort to make Kanto a bit more interesting.

The code for teaching Pikachu is very buggy right now and someone should check it out. I usually throw code out if it isn't functional but this seems extremely close to working.