King of the Crypt
Dokuroar, in all of its skeletal glory, has been added to the bottom of Lavender Crypt as a static encounter. It doesn't move when you interact with it, though...at first.

Also, gave Dokuroar a 100% chance to hold a Thick Club (it's MADE of bones, come on), a Snorlax-level catch rate, the Slow EXP group, a decent exp yield, and a very, VERY slight stat update.