New dex entries using DP myth docs
Piloswine and Lapras pending; seem weak.
Octillery story seems to lewd even for a game that has decensored much of the original content.

Rapidash and Tauros entries give credence to featuring them around Ruddy Town with Coatl and Xatu.

Typhlosion entry allows for connections to the mujina (even if not wholly intentional, seemingly), which I quite enjoy. However, the dex entry is unsatisfactory.