Ajawalcan is here! Also, Inoshika buffs.
This adds Ajawalcan's new name, using Mayan grammar and vocabulary resources to produce something truly befitting of the Kukulcan Pokemon. Before, we conflated Aztec and Mayan theology and have now chosen to squarely focus on the Mayan side of things, with Ruddy Ruins scheduled to be based on Chichen Itza.

* Bolles, D. & Bolles, A. (1996). A Grammar of the Yucatecan Mayan Language. Revised ed. Lancaster, California: Labyrinthos Press. ISBN 0-911437-49-5. Available at: http://www.famsi.org/reports/96072/grammar/index.html
* Boot, E. (2009). The Updated Preliminary Classic Maya - English, English - Classic Maya Vocabulary of Hieroglyphic Readings. Updated edition. Mesoweb Resources. Available at: https://www.mesoweb.com/resources/vocabulary/Vocabulary-2009.01.pdf