ITEMFIX: I hate working with WRAM
CHANGES ARE COMMENTED IN WRAM.ASM, IF ANYTHING BREAKS, CHECK THERE.

Not thoroughly tested but I've yeeted some unused shit in WRAM to be able to fit Misc WRAM 1 back in the right bank.
Side note but according to pret if a ds has a blank space above and below it then it should be unused? May be wise to double check this